backend/requirements.txt
django==4.2.6
djangorestframework==3.14.0
pymongo==4.6.0
python-decouple==3.8
requests